chore: move isValidUUID out of UUIDValidationInterceptor
