+++ skill-tree/src/main/java/com/RDS/skilltree/utils/UUIDValidationInterceptor.java
@@ -3,7 +3,6 @@ package com.RDS.skilltree.utils;
 import com.RDS.skilltree.Exceptions.InvalidParameterException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.regex.Pattern;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
@@ -13,9 +12,6 @@ import org.springframework.web.servlet.HandlerInterceptor;
 public class UUIDValidationInterceptor implements HandlerInterceptor {
 
     private static final Logger logger = LoggerFactory.getLogger(UUIDValidationInterceptor.class);
-    private static final Pattern UUID_REGEX =
-            Pattern.compile(
-                    "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$");
 
     @Override
     public boolean preHandle(
@@ -23,19 +19,14 @@ public class UUIDValidationInterceptor implements HandlerInterceptor {
         String skillID = request.getParameter("skillID");
         String userID = request.getParameter("userID");
 
-        if (skillID != null && !skillID.isEmpty() && !isValidUUID(skillID)) {
+        if (skillID != null && !skillID.isEmpty() && !CommonUtils.isValidUUID(skillID)) {
             throw new InvalidParameterException("skillID", "Invalid UUID format");
         }
 
-        if (userID != null && !userID.isEmpty() && !isValidUUID(userID)) {
+        if (userID != null && !userID.isEmpty() && !CommonUtils.isValidUUID(userID)) {
             throw new InvalidParameterException("userID", "Invalid UUID format");
         }
 
         return true;
-    }
-
-    public static boolean isValidUUID(String uuidString) {
-
-        return UUID_REGEX.matcher(uuidString).matches();
     }
 }
